refactor: createWorkoutHandler, fix getUserAttr

+++ api/handlers/auth.go
@@ -24,7 +24,7 @@ type registerUserPayload struct {
 	Goal       string  `json:"goal"        validate:"required,oneof=lose gain maintain"`
 	WeightGoal float32 `json:"weight_goal"`
 	Weight     float32 `json:"weight"`
-	Age        int     `json:"age"`
+	Age        int64   `json:"age"`
 }
 
 type TokenResponse struct {
